+++ src/protocol.rs
@@ -18,9 +18,6 @@ impl Protocol {
     pub fn to_server(config: &Config, protocols: Vec<Self>) -> Server {
         let mut server = ServerBuilder::new();
 
-        server = config.durations.apply(server);
-        server = config.hooks.apply(server);
-
         let protocols = if protocols.is_empty() {
             vec![
                 #[cfg(feature = "tcp-binder")]
@@ -34,10 +31,11 @@ impl Protocol {
             match protocol {
                 #[cfg(feature = "tcp-binder")]
                 Protocol::Tcp => {
-                    if let Some(ref config) = config.tcp {
-                        server = server.with_binder(TcpBind::new(&config.host, config.port));
-                        server = config.durations.apply(server);
-                        server = config.hooks.apply(server);
+                    if let Some(ref tcp_config) = config.tcp {
+                        let binder = TcpBind::new(&tcp_config.host, tcp_config.port);
+                        server = server.with_binder(binder);
+                        server = tcp_config.durations.apply(config, server);
+                        server = tcp_config.hooks.apply(config, server);
                     }
                 }
                 Protocol::None => (),